Add caching functionality
Up to now the folksonomy table was created on each and every access.
The more entries you had the longer it took to create this table.
With this change the plugin looks for an existing cache file which can
be created via the pyblosxom-cmd command and uses it instead. For more
information read revert to the documentation inside the plugin.